extraí a função dateFormat para outro arquivo
